refactor: pull fetching logic out of form
